update 2.6.23
Models updates to create a unique realtionship on the Diet Restricitons table.
This is being done so that when the filter query is built is can be sure
that all relationships between a given Id and given type of diet restriction
is unique.

The raw query will be a:
SELECT count(diet_restriciton), item_id
FROM (
    SELECT * FROM menu_diet_restriction WHERE
        [set of restrictions from GET request])
    AS restrictionTable
GROUPED BY
 item_id
HAVING
  count(diet_restriction) = [ count ofr restrictions from GET]
ORDER BY
  count(diet_restriction);

this query is hinged on the implication that everytime a item_id is
listed it is the only time it appears for a given restriction so a simple
count of it's occurences will be used to filter any that don't appear for all
given restrictions.