Update 2.6.24
Updated the pseudo code with python code. Subquery is produced and used in a
query to the Diet_Restrictions table that will filter out only the items that
satisfy all the given restrictions. Used python code to build a subquery
instead of filtering a return in python to limit calls to the database and
to limited wasted data being pulled.

The next update will be to filter the Items table query using the returned list
of IDs from the Diet_Restrictions query. This will mock a left outer join with
out having to bring the added data into the list passed to the render.

Hopefully this will limit time spent querying the table and only pull data
needed.