Update 2.6.21
Moved the sql builde to a seperate function for ease of reading and building.
At this point the sql is build using Django's ORM methods. With more testing
and protections this may change to perform raw queries in order to cut down on
redundant code. As of now it has a set list of OR statments to avoid query
issues. With more refactoring this may be able to be replicated with less
code using raw sql queries and moving it to a function will make it easier
to preserv functionality with testing.